requirements.txt
requests
discord.py
schedule
asyncio
json